feat: created dynamic data
