+++ config/database.json
@@ -1,1 +1,1 @@
-[]
+[{"id":"8852b3f8-ee6f-46af-ab61-3c45702e6bf9","name":"Default 1","description":"Default Cube Description 1","imageUrl":"https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg","difficulty":1},{"id":"6d52d67c-8ad2-4a7f-8105-68dd30fde8e7","name":"Default 2","description":"Default Cube Description 2","imageUrl":"https://thingsidesire.com/wp-content/uploads/2018/06/Eco-Dark-Rubik%E2%80%99s-Cube2.jpg","difficulty":2},{"id":"bc39af89-89ef-4b21-a756-3812ec08a861","name":"Default 3","description":"Default Cube Description 3","imageUrl":"https://images-na.ssl-images-amazon.com/images/I/61izOzq%2BBAL._SY355_.jpg","difficulty":3},{"id":"3deeb477-9544-4666-b853-0a7fbf72a99a","name":"Default 4","description":"Default Cube Description 4","imageUrl":"https://images-na.ssl-images-amazon.com/images/I/61HpQqVQ37L._SY355_.jpg","difficulty":4}]
