+++ controllers/create-cube.js
@@ -1,7 +1,13 @@
 const Cube = require('../models/cube');
 
-const expampleCube = new Cube('Default', 'Default Cube Description', 'google.com', 1);
+// const expampleCube1 = new Cube('Default 1', 'Default Cube Description 1', 'https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg', 1);
+// const expampleCube2 = new Cube('Default 2', 'Default Cube Description 2', 'https://thingsidesire.com/wp-content/uploads/2018/06/Eco-Dark-Rubik%E2%80%99s-Cube2.jpg', 2);
+// const expampleCube3 = new Cube('Default 3', 'Default Cube Description 3', 'https://images-na.ssl-images-amazon.com/images/I/61izOzq%2BBAL._SY355_.jpg', 3);
+// const expampleCube4 = new Cube('Default 4', 'Default Cube Description 4', 'https://images-na.ssl-images-amazon.com/images/I/61HpQqVQ37L._SY355_.jpg', 4);
 
-console.log(expampleCube);
+// console.log(expampleCube1);
 
-expampleCube.save();
+// expampleCube1.save();
+// expampleCube2.save();
+// expampleCube3.save();
+// expampleCube4.save();
