+++ routes/index.js
@@ -1,11 +1,15 @@
 // TODO: Require Controllers...
 const { Router } = require('express');
 const router = Router();
+const { getAllCubes, getCube } = require('../controllers/database');
 
 router.get('/', (req, res) => {
-    res.render('index', {
-        title: 'Cube Workshop'
-    });
+    getAllCubes((cubes) => {
+        res.render('index', {
+            title: 'Cube Workshop',
+            cubes
+        });
+    })
 });
 
 router.get('/about', (req, res) => {
@@ -21,10 +25,13 @@ router.get('/create', (req, res) => {
 });
 
 router.get('/details/:id', (req, res) => {
-    res.render('details', {
-        title: 'Cube Details | Cube Workshop'
+    getCube(req.params.id, (cube) => {
+        res.render('details', {
+            title: 'Cube Details | Cube Workshop',
+            ...cube
+        });
     });
-});
+ });
 
 router.get('*', (req, res) => {
     res.render('404', {
